refactor: application interfaces
