+++ src/application/interfaces/scanner.interface.ts
