+++ src/application/services/app.service.ts
@@ -7,7 +7,7 @@ import { progressBarExecutableMetadata } from "../utils/progress-bar-executable-
 import { ExecutableService } from "./executable-service";
 import { ExternalFileService } from "./external-file.service";
 import { ProcessesScanner, ScheduledTasksScanner, ServicesScanner } from "./scanners";
-import { IScanner } from "./scanners/scanner.interface";
+import { IScanner } from "../interfaces/scanner.interface";
 import { TOptions } from "./types";
 
 @Service()
