+++ src/application/services/scanners/processes.scanner.ts
@@ -4,7 +4,7 @@ import { IExecutable } from "@domain/interfaces/executable.interface";
 import { basename } from "node:path";
 import { Inject, Service } from "typedi";
 import { Powershell } from "../../utils/powershell";
-import { IScanner } from "./scanner.interface";
+import { IScanner } from "../../interfaces/scanner.interface";
 
 @Service()
 export class ProcessesScanner implements IScanner {
