+++ src/application/services/scanners/services.scranner.ts
@@ -4,7 +4,7 @@ import { ExecutableSource } from "@domain/enum/executable-source.enum";
 import { IExecutable } from "@domain/interfaces/executable.interface";
 import { Service } from "typedi";
 import { Powershell } from "../../utils/powershell";
-import { IScanner } from "./scanner.interface";
+import { IScanner } from "../../interfaces/scanner.interface";
 
 @Service()
 export class ServicesScanner implements IScanner {
